feat: separate subscription hooks into /hooks
